+++ src/hooks/LevelCommentDelegate.cpp
@@ -1,7 +1,0 @@
-#include "LevelCommentDelegate.hpp"
-
-using namespace geode::prelude;
-
-void CKLevelCommentDelegate::loadCommentsFinished(CCArray* comments, const char* key) {
-    log::info("Comments loaded for key: {}", key);
-}