+++ changelog.md
@@ -1,2 +1,4 @@
-# 1.0.0
-- Edit this file to change your mod's changelog.
+# Comment Karma Changelog
+
+## 1.0.0
+* Initial release, showing karma!